move `FieldName` and `FieldNames` to `field_names.rs`
