+++ vortex-dtype/src/lib.rs
@@ -9,16 +9,6 @@
 //! This crate contains the core logical type system for Vortex, including the definition of data types,
 //! and (optionally) logic for their serialization and deserialization.
 
-pub use decimal::*;
-pub use dtype::*;
-pub use extension::*;
-pub use field::*;
-pub use field_mask::*;
-pub use half;
-pub use nullability::*;
-pub use ptype::*;
-pub use struct_::*;
-
 #[cfg(feature = "arbitrary")]
 mod arbitrary;
 #[cfg(feature = "arrow")]
@@ -29,10 +19,22 @@ mod dtype;
 mod extension;
 mod field;
 mod field_mask;
+mod field_names;
 mod nullability;
 mod ptype;
 mod serde;
 mod struct_;
+
+pub use decimal::*;
+pub use dtype::*;
+pub use extension::*;
+pub use field::*;
+pub use field_mask::*;
+pub use field_names::*;
+pub use half;
+pub use nullability::*;
+pub use ptype::*;
+pub use struct_::*;
 
 pub mod proto {
     //! Protocol buffer representations for DTypes
